refactor: refrigerator 관련 컴포넌트, 페이지 디렉토리 구조 정리
