+++ src/components/refrigerator/main/RefrigeratorSVG.css
