+++ src/components/refrigerator/main/RefrigeratorSVG.js
