+++ src/components/refrigerator/main/RefrigeratorSVGv2.css
