+++ src/components/refrigerator/main/RefrigeratorSVGv2.js
