+++ src/components/refrigerator/popup/IngredientDetailContent.css
