+++ src/pages/refrigerator/list/IngredientDetailList.css
